requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0.1
django-allauth==0.57.0
django-cors-headers==4.3.1
django-filter==23.5
django-push-notifications==3.0.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
pycparser==2.21
pyfcm==1.5.4
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
razorpay
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0